Remove unuse file and correct some method name
