+++ core/data/coercion_test.go
@@ -97,3 +97,13 @@ func TestCoerceToBoolean(t *testing.T) {
 	cval, _ = CoerceToBoolean(valNil)
 	assert.Equal(t, false, cval, "not equal")
 }
+
+
+func TestUnmarshComplexObject(t *testing.T) {
+	complexStr := `{"metadata":"this is metdata string","value":""}`
+	complexObject, err := CoerceToComplexObject(complexStr)
+	assert.Nil(t, err)
+	assert.NotNil(t, complexObject)
+	assert.NotEqual(t, "", complexObject.Value)
+}
+
